feat: paste screenshots into the report dialog (JUN-213)
Pasting an image (CleanShot and friends) attaches it through the same
import pipeline as drag-drop, with the composer's clipboard extraction
moved to lib/clipboard-files so both surfaces share it. Text pastes fall
through to the textarea untouched.
